Remove DeathMessages listener due to low-quality implementation relying on string parsing
This implementation was written a long time ago (early in my programming journey) and relies heavily on parsing death message strings. This approach is brittle, prone to breaking with future Minecraft updates, and difficult to test. It also doesn't account for potential issues with translations. I will most likely revisit and rewrite this feature with a more robust approach in the future.